[#175] Data module refactor (#176)
* Initial rewrite

* Added alias class in persistence xml

* Updated rest project to correctly use new data module

* Fixe data sql problem

* Updated owner table

* Fixed stupid mini error

* More updates and fixes to data module

* Fixed commands that abused upsert

+++ environment/sql/tables/guild_setting.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `guild_setting` (
-  `id` int(11) PRIMARY KEY AUTO_INCREMENT,
+  `id` bigint(11) PRIMARY KEY AUTO_INCREMENT,
   `guildFK` bigint(11) NOT NULL,
   `name` varchar(50) NOT NULL,
   `value` varchar(200) NOT NULL
