+++ environment/sql/tables/metric.sql
@@ -1,8 +1,8 @@
 CREATE TABLE `metric` (
-  `ìd` int(11) PRIMARY KEY AUTO_INCREMENT,
+  `ìd` bigint(11) PRIMARY KEY AUTO_INCREMENT,
   `name` varchar(30) NOT NULL,
   `date` TIMESTAMP NOT NULL,
-  `commandFK` int(11) NULL,
+  `commandFK` bigint(11) NULL,
   `guildFK` bigint(11) NULL,
   `userFK` bigint(11) NULL,
   `value` bigint(11) NULL,
